endpoint for getting fooclespots for business accounts
