+++ src/main/java/facades/FoocleBusinessFacade.java
@@ -60,65 +60,6 @@ public class FoocleBusinessFacade {
         return new FoocleBusinessDTO(foocleBusiness);
     }
 
-    public FoocleSpotDTO createFoocleSpot(long businessAccountID, String address, String city, String zipCode, String country) {
-        Location location = new Location(address, city, zipCode, country);
-
-        FoocleSpot spot =  executeWithClose(em -> {
-            TypedQuery<BusinessAccount> query = em.createQuery("SELECT f FROM BusinessAccount f WHERE f.id = :id", BusinessAccount.class);
-            query.setParameter("id", businessAccountID);
-            TypedQuery<FoocleBusiness> query2 = em.createQuery("SELECT f FROM FoocleBusiness f WHERE f.id = :cvr", FoocleBusiness.class);
-            query2.setParameter("cvr", query.getResultList().get(0).getCvr().getId());
-            return new FoocleSpot(query.getResultList().get(0), query2.getResultList().get(0), location);
-        });
-        List<Location> responseLocation = executeWithClose(em -> {
-            TypedQuery<Location> query = em.createQuery("SELECT l FROM Location l " +
-                    "WHERE l.address = :address AND l.city = :city AND l.country = :country", Location.class);
-            query.setParameter("address", address);
-            query.setParameter("city", city);
-            query.setParameter("country", country);
-            return query.getResultList();
-        });
-
-        executeInsideTransaction(em -> {
-            if (responseLocation.isEmpty()) {
-                em.persist(location);
-            } else {
-                spot.setLocation(responseLocation.get(0));
-            }
-            em.persist(spot);
-        });
-        return new FoocleSpotDTO(spot);
-    }
-    public List<FoocleSpotDTO> getFoocleSpotsForCVR(long businessAccountID) {
-        List<FoocleSpot> spots =  executeWithClose(em -> {
-            TypedQuery<BusinessAccount> query = em.createQuery("SELECT f FROM BusinessAccount f WHERE f.id = :id", BusinessAccount.class);
-            query.setParameter("id", businessAccountID);
-            TypedQuery<FoocleSpot> query2 = em.createQuery("SELECT f FROM FoocleSpot f WHERE f.cvr = :cvr", FoocleSpot.class);
-            query2.setParameter("cvr", query.getResultList().get(0).getCvr().getId());
-            return query2.getResultList();
-        });
-
-        List<FoocleSpotDTO> result = new ArrayList<>();
-        for (FoocleSpot spot:spots) {
-            result.add(new FoocleSpotDTO(spot));
-        }
-
-        return result;
-    }
-
-
-    public SpotMenuDTO createSpotMenu(String description, String pictures, String foodPrefences, LocalDateTime pickupTimeFrom, LocalDateTime pickupTimeTo, long fooclespotID) {
-        FoocleSpot foocleSpot = executeWithClose(em -> em.find(FoocleSpot.class, fooclespotID));
-        SpotMenu spotMenu = new SpotMenu(description, pictures, foodPrefences, pickupTimeFrom, pickupTimeTo, foocleSpot);
-
-        executeInsideTransaction(em -> {
-            em.persist(spotMenu);
-        });
-
-        return new SpotMenuDTO(spotMenu);
-    }
-
-
 
     private <R> R executeWithClose(Function<EntityManager, R> action) {
         EntityManager em = emf.createEntityManager();
